cmd/snap: split out wait, and always set up the sigint handler
