+++ cmd/snap/cmd_snap_op.go
@@ -23,19 +23,15 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"os/signal"
 	"path/filepath"
 	"sort"
 	"strings"
-	"syscall"
-	"time"
 
 	"github.com/jessevdk/go-flags"
 
 	"github.com/snapcore/snapd/client"
 	"github.com/snapcore/snapd/i18n"
 	"github.com/snapcore/snapd/osutil"
-	"github.com/snapcore/snapd/progress"
 )
 
 func lastLogStr(logs []string) string {
@@ -43,104 +39,6 @@ func lastLogStr(logs []string) string {
 		return ""
 	}
 	return logs[len(logs)-1]
-}
-
-var (
-	maxGoneTime = 5 * time.Second
-	pollTime    = 100 * time.Millisecond
-)
-
-type waitMixin struct {
-	NoWait bool `long:"no-wait"`
-}
-
-// TODO: use waitMixin outside of cmd_snap_op.go
-
-var waitDescs = mixinDescs{
-	"no-wait": i18n.G("Do not wait for the operation to finish but just print the change id."),
-}
-
-var noWait = errors.New("no wait for op")
-
-func (wmx waitMixin) wait(cli *client.Client, id string) (*client.Change, error) {
-	if wmx.NoWait {
-		fmt.Fprintf(Stdout, "%s\n", id)
-		return nil, noWait
-	}
-	pb := progress.MakeProgressBar()
-	defer func() {
-		pb.Finished()
-	}()
-
-	tMax := time.Time{}
-
-	var lastID string
-	lastLog := map[string]string{}
-	for {
-		chg, err := cli.Change(id)
-		if err != nil {
-			// a client.Error means we were able to communicate with
-			// the server (got an answer)
-			if e, ok := err.(*client.Error); ok {
-				return nil, e
-			}
-
-			// an non-client error here means the server most
-			// likely went away
-			// XXX: it actually can be a bunch of other things; fix client to expose it better
-			now := time.Now()
-			if tMax.IsZero() {
-				tMax = now.Add(maxGoneTime)
-			}
-			if now.After(tMax) {
-				return nil, err
-			}
-			pb.Spin(i18n.G("Waiting for server to restart"))
-			time.Sleep(pollTime)
-			continue
-		}
-		if !tMax.IsZero() {
-			pb.Finished()
-			tMax = time.Time{}
-		}
-
-		for _, t := range chg.Tasks {
-			switch {
-			case t.Status != "Doing":
-				continue
-			case t.Progress.Total == 1:
-				pb.Spin(t.Summary)
-				nowLog := lastLogStr(t.Log)
-				if lastLog[t.ID] != nowLog {
-					pb.Notify(nowLog)
-					lastLog[t.ID] = nowLog
-				}
-			case t.ID == lastID:
-				pb.Set(float64(t.Progress.Done))
-			default:
-				pb.Start(t.Summary, float64(t.Progress.Total))
-				lastID = t.ID
-			}
-			break
-		}
-
-		if chg.Ready {
-			if chg.Status == "Done" {
-				return chg, nil
-			}
-
-			if chg.Err != "" {
-				return chg, errors.New(chg.Err)
-			}
-
-			return nil, fmt.Errorf(i18n.G("change finished in status %q with no error message"), chg.Status)
-		}
-
-		// note this very purposely is not a ticker; we want
-		// to sleep 100ms between calls, not call once every
-		// 100ms.
-		time.Sleep(pollTime)
-	}
 }
 
 var (
@@ -431,20 +329,6 @@ type cmdInstall struct {
 	} `positional-args:"yes" required:"yes"`
 }
 
-func setupAbortHandler(changeId string) {
-	// Intercept sigint
-	c := make(chan os.Signal, 2)
-	signal.Notify(c, syscall.SIGINT)
-	go func() {
-		<-c
-		cli := Client()
-		_, err := cli.Abort(changeId)
-		if err != nil {
-			fmt.Fprintf(Stderr, err.Error()+"\n")
-		}
-	}()
-}
-
 func (x *cmdInstall) installOne(name string, opts *client.SnapOptions) error {
 	var err error
 	var installFromFile bool
@@ -465,8 +349,6 @@ func (x *cmdInstall) installOne(name string, opts *client.SnapOptions) error {
 		fmt.Fprintln(Stderr, msg)
 		return nil
 	}
-
-	setupAbortHandler(changeID)
 
 	chg, err := x.wait(cli, changeID)
 	if err != nil {
@@ -511,8 +393,6 @@ func (x *cmdInstall) installMany(names []string, opts *client.SnapOptions) error
 		fmt.Fprintln(Stderr, msg)
 		return nil
 	}
-
-	setupAbortHandler(changeID)
 
 	chg, err := x.wait(cli, changeID)
 	if err != nil {
